Revert "Add new Docker SDK-based client (#2134)"
This reverts commit cb41bfbf89e704c50e255429a4861305fe209f45.
