+++ package.json
@@ -2652,7 +2652,6 @@
         "webpack-cli": "^3.3.12"
     },
     "dependencies": {
-        "@docker/sdk": "^0.1.7",
         "adal-node": "^0.2.1",
         "azure-arm-containerregistry": "^5.1.0",
         "azure-arm-website": "^5.7.0",
