+++ src/commands/containers/attachShellContainer.ts
@@ -5,7 +5,6 @@
 
 import * as vscode from 'vscode';
 import { IActionContext } from 'vscode-azureextensionui';
-import { DockerOSType } from '../../docker/Common';
 import { ext } from '../../extensionVariables';
 import { localize } from '../../localize';
 import { ContainerTreeItem } from '../../tree/containers/ContainerTreeItem';
@@ -21,15 +20,7 @@ export async function attachShellContainer(context: IActionContext, node?: Conta
         });
     }
 
-    let osType: DockerOSType;
-    try {
-        // TODO: get OS type from container instead of from system
-        osType = await getDockerOSType(context);
-    } catch {
-        // Assume linux
-        osType = 'linux';
-    }
-
+    let osType = await getDockerOSType(context);
     context.telemetry.properties.dockerOSType = osType;
 
     let shellCommand: string;
