+++ src/docker/Containers.ts
@@ -12,12 +12,6 @@ export interface DockerPort {
     readonly Type?: string;
 }
 
-// Ports from inspect have a different shape entirely
-export interface InspectionPort {
-    readonly HostIp?: string;
-    readonly HostPort?: string;
-}
-
 export interface DockerContainer extends DockerObject {
     readonly State: string;
     readonly Status: string;
@@ -38,7 +32,10 @@ export interface DockerContainerInspection extends DockerObject {
     };
     readonly NetworkSettings?: {
         readonly Ports?: {
-            readonly [portAndProtocol: string]: InspectionPort[];
+            readonly [portAndProtocol: string]: {
+                readonly HostIp?: string;
+                readonly HostPort?: string;
+            }[];
         };
     };
 }
