+++ src/docker/ContextManager.ts
@@ -46,7 +46,6 @@ export interface ContextManager {
     readonly onContextChanged: Event<DockerContext>;
     refresh(): Promise<void>;
     getContexts(): Promise<DockerContext[]>;
-    getCurrentContext(): Promise<DockerContext>;
 
     inspect(actionContext: IActionContext, contextName: string): Promise<DockerContextInspection>;
     use(actionContext: IActionContext, contextName: string): Promise<void>;
@@ -97,9 +96,8 @@ export class DockerContextManager implements ContextManager, Disposable {
             this.refreshing = true;
 
             this.contextsCache.clear();
-
-            // Because the cache is cleared, this will load all the contexts before returning the current one
-            const currentContext = await this.getCurrentContext();
+            const contexts = await this.contextsCache.getValue();
+            const currentContext = contexts.find(c => c.Current);
 
             void ext.dockerClient?.dispose();
 
@@ -128,11 +126,6 @@ export class DockerContextManager implements ContextManager, Disposable {
 
     public async getContexts(): Promise<DockerContext[]> {
         return this.contextsCache.getValue();
-    }
-
-    public async getCurrentContext(): Promise<DockerContext> {
-        const contexts = await this.getContexts();
-        return contexts.find(c => c.Current);
     }
 
     public async inspect(actionContext: IActionContext, contextName: string): Promise<DockerContextInspection> {
