+++ src/docker/DockerServeClient/DockerServeClient.ts
@@ -3,11 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Containers as ContainersClient } from '@docker/sdk';
-import { DeleteRequest, InspectRequest, InspectResponse, ListRequest, ListResponse } from '@docker/sdk/containers';
 import { CancellationToken } from 'vscode';
 import { IActionContext } from 'vscode-azureextensionui';
-import { localize } from '../../localize';
 import { DockerInfo, PruneResult } from '../Common';
 import { DockerContainer, DockerContainerInspection } from '../Containers';
 import { ContextChangeCancelClient } from '../ContextChangeCancelClient';
@@ -16,60 +13,22 @@ import { DockerImage, DockerImageInspection } from '../Images';
 import { DockerNetwork, DockerNetworkInspection, DriverType } from '../Networks';
 import { NotSupportedError } from '../NotSupportedError';
 import { DockerVolume, DockerVolumeInspection } from '../Volumes';
-import { containerPortsToInspectionPorts, containerToDockerContainer } from './DockerServeUtils';
-
-// 20 s timeout for all calls (enough time for any call, but short enough to be UX-reasonable)
-const dockerServeCallTimeout = 20 * 1000;
 
 export class DockerServeClient extends ContextChangeCancelClient implements DockerApiClient {
-    private readonly containersClient: ContainersClient;
-
-    public constructor() {
-        super();
-        this.containersClient = new ContainersClient();
-    }
-
-    public dispose(): void {
-        super.dispose();
-        void this.containersClient?.close();
-    }
-
     public async info(context: IActionContext, token?: CancellationToken): Promise<DockerInfo> {
         throw new NotSupportedError(context);
     }
 
     public async getContainers(context: IActionContext, token?: CancellationToken): Promise<DockerContainer[]> {
-        const request = new ListRequest()
-            .setAll(true);
-
-        const response: ListResponse = await this.promisify(context, this.containersClient, this.containersClient.list, request, token);
-        const result = response.getContainersList();
-
-        return result.map(c => containerToDockerContainer(c.toObject()));
-    }
-
-    public async inspectContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<DockerContainerInspection> {
-        const request = new InspectRequest()
-            .setId(ref);
-
-        const response: InspectResponse = await this.promisify(context, this.containersClient, this.containersClient.inspect, request, token);
-        const container = containerToDockerContainer(response.toObject().container);
-
-        if (!container) {
-            throw new Error(localize('vscode-docker.dockerServeClient.noContainer', 'No container with name \'{0}\' was found.', ref));
-        }
-
-        return {
-            ...container,
-            NetworkSettings: {
-                Ports: containerPortsToInspectionPorts(container),
-            },
-        };
+        throw new NotSupportedError(context);
     }
 
     // #region Not supported by the Docker SDK yet
+    public async inspectContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<DockerContainerInspection> {
+        throw new NotSupportedError(context);
+    }
+
     public async getContainerLogs(context: IActionContext, ref: string, token?: CancellationToken): Promise<NodeJS.ReadableStream> {
-        // Supported by SDK, but used only for debugging which will not work in ACI, and complicated to implement
         throw new NotSupportedError(context);
     }
 
@@ -84,19 +43,14 @@ export class DockerServeClient extends ContextChangeCancelClient implements Dock
     public async restartContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<void> {
         throw new NotSupportedError(context);
     }
-
-    public async stopContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<void> {
-        // Supported by SDK, but is not really the same thing; containers in ACI must stop/start as a group
-        throw new NotSupportedError(context);
-    }
     // #endregion Not supported by the Docker SDK yet
 
-    public async removeContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<void> {
-        const request = new DeleteRequest()
-            .setId(ref)
-            .setForce(true);
+    public async stopContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<void> {
+        throw new NotSupportedError(context);
+    }
 
-        await this.promisify(context, this.containersClient, this.containersClient.delete, request, token)
+    public async removeContainer(context: IActionContext, ref: string, token?: CancellationToken): Promise<void> {
+        throw new NotSupportedError(context);
     }
 
     // #region Not supported by the Docker SDK yet
@@ -156,28 +110,4 @@ export class DockerServeClient extends ContextChangeCancelClient implements Dock
         throw new NotSupportedError(context);
     }
     // #endregion Not supported by the Docker SDK yet
-
-    private async promisify<TRequest, TResponse>(
-        context: IActionContext,
-        thisArg: unknown,
-        clientCallback: (request: TRequest, callback: (err: unknown, response: TResponse) => void) => unknown,
-        request: TRequest,
-        token?: CancellationToken): Promise<TResponse> {
-
-        const callPromise: Promise<TResponse> = new Promise((resolve, reject) => {
-            try {
-                clientCallback.call(thisArg, request, (err, response) => {
-                    if (err) {
-                        reject(err);
-                    }
-
-                    resolve(response);
-                });
-            } catch (err) {
-                reject(err);
-            }
-        });
-
-        return this.withTimeoutAndCancellations(context, async () => callPromise, dockerServeCallTimeout, token);
-    }
 }
