+++ src/docker/DockerodeApiClient/DockerodeApiClient.ts
@@ -15,9 +15,10 @@ import { DockerApiClient } from '../DockerApiClient';
 import { DockerImage, DockerImageInspection } from '../Images';
 import { DockerNetwork, DockerNetworkInspection, DriverType } from '../Networks';
 import { DockerVolume, DockerVolumeInspection } from '../Volumes';
-import { getContainerName, getFullTagFromDigest, refreshDockerode } from './DockerodeUtils';
+import { getContainerName, getFullTagFromDigest } from './DockerodeUtils';
+import { refreshDockerode } from './refreshDockerode';
 
-// 20 s timeout for all calls (enough time for any call, but short enough to be UX-reasonable)
+// 20 s timeout for all calls (enough time for a possible Dockerode refresh + the call, but short enough to be UX-reasonable)
 const dockerodeCallTimeout = 20 * 1000;
 
 export class DockerodeApiClient extends ContextChangeCancelClient implements DockerApiClient {
